RestExercise : log endpoint extended pagination, findall does not work however, findallbydatacontaining works

+++ week24/rest1/src/main/java/com/greenfoxacademy/restfirst/model/EntryCounter.java
@@ -1,26 +1,28 @@
 package com.greenfoxacademy.restfirst.model;
 
 
+import org.springframework.data.domain.Page;
+
 import java.util.List;
 
 public class EntryCounter {
 
   private List<Log> entries;
-  private int entry_count;
+  private long entry_count;
 
   public EntryCounter() {
   }
 
-  public EntryCounter(List<Log> logs, int entry_count) {
+  public EntryCounter(List<Log> logs, long entry_count) {
     this.entries = logs;
     this.entry_count = entry_count;
   }
 
-  public int getEntry_count() {
+  public long getEntry_count() {
     return entry_count;
   }
 
-  public void setEntry_count(int entry_count) {
+  public void setEntry_count(long entry_count) {
     this.entry_count = entry_count;
   }
